ft: update copyable component
