+++ design-system/react/src/defs.ts
@@ -37,6 +37,7 @@ import type { ButtonDef } from './components/Button/defs';
 import type { ButtonGroupDef } from './components/ButtonGroup/defs';
 import type { CardListDef, CardListItemDef } from './components/CardList/defs';
 import type { CheckboxDef } from './components/Checkbox/defs';
+import type { CopyableDef } from './components/Copyable/defs';
 import type { HeadingDef } from './components/Heading/defs';
 import type { HelperIconDef } from './components/HelperIcon/defs';
 import type { IconDef } from './components/Icon/defs';
@@ -182,6 +183,7 @@ export type StoreDefs = {
   CardList: CardListDef;
   CardListItem: CardListItemDef;
   Checkbox: CheckboxDef;
+  Copyable: CopyableDef;
   Container: ContainerDef;
   Image: ImageDef;
   Flex: FlexDef;
